+++ CHANGES.rst
@@ -8,6 +8,7 @@ Development
 - Convert to Python3
 - Add setup.py, README.rst and CHANGES.rst
 - Provide ``extract_sites.sh`` through ``apicast beeflight stations``
+- Output beeflight forecast in JSON or table format
 
 
 2018-04-13 0.1.0